# docs/openapi/openapi.v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKDY RAG Server API",
    "description": "HTTP API for the SKDY RAG Server. Mounted under `/api/v1`.\n\n**Contract**: the OpenAPI schema is the single source of truth for the Web UI; client types are generated from it.\n\n**Error envelope**: every 4xx/5xx returns `{\"error\": {\"code\": \"...\", \"message\": \"...\", \"request_id\": \"...\", \"details\": {}}}`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/documents/{document_id}/chunks/{chunk_id}/tag-suggestions": {
      "post": {
        "tags": [
          "knowledge-enrichment"
        ],
        "summary": "Generate Tag Suggestions",
        "operationId": "generate_tag_suggestions_api_v1_documents__document_id__chunks__chunk_id__tag_suggestions_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Generate Tag Suggestions Api V1 Documents  Document Id  Chunks  Chunk Id  Tag Suggestions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/tag-suggestions": {
      "get": {
        "tags": [
          "knowledge-enrichment"
        ],
        "summary": "List Tag Suggestions",
        "operationId": "list_tag_suggestions_api_v1_documents__document_id__tag_suggestions_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "revision_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Tag Suggestions Api V1 Documents  Document Id  Tag Suggestions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tag-suggestions/{suggestion_id}/review": {
      "post": {
        "tags": [
          "knowledge-enrichment"
        ],
        "summary": "Review Tag Suggestion",
        "operationId": "review_tag_suggestion_api_v1_tag_suggestions__suggestion_id__review_post",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Suggestion Id"
            }
          },
          {
            "name": "X-Actor-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "X-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Tag Suggestion Api V1 Tag Suggestions  Suggestion Id  Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/revisions/{revision_id}/derived-artifacts": {
      "get": {
        "tags": [
          "knowledge-enrichment"
        ],
        "summary": "List Derived Artifacts",
        "operationId": "list_derived_artifacts_api_v1_revisions__revision_id__derived_artifacts_get",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Derived Artifacts Api V1 Revisions  Revision Id  Derived Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "knowledge-enrichment"
        ],
        "summary": "Delete Derived Artifacts",
        "operationId": "delete_derived_artifacts_api_v1_revisions__revision_id__derived_artifacts_delete",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Delete Derived Artifacts Api V1 Revisions  Revision Id  Derived Artifacts Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/derived-artifacts/{artifact_id}": {
      "patch": {
        "tags": [
          "knowledge-enrichment"
        ],
        "summary": "Set Derived Artifact Enabled",
        "operationId": "set_derived_artifact_enabled_api_v1_derived_artifacts__artifact_id__patch",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerivedEnabledRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Derived Artifact Enabled Api V1 Derived Artifacts  Artifact Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DerivedEnabledRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "DerivedEnabledRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewRequest": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "resolved_tag_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Tag Id"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "ReviewRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}